Throw an exception on non-skipped unresolved dependency.
Currently this would cause an circular dependency exception if a following package referred to the package that has the unresolved dependency.

I also had to alter the uninstall and install walkers to call a new method in dependents walker so they can correctly skip a failed dependency.

The install walker collects failures and alerts later, so this is OK